Use display information from layers trace on assertions
Displays were added to the SF trace. Parse them on flicker and use during navBarLayerRotatesAndScales and statusBarLayerRotatesScales assertion instead of relying on the test spec rotation.

Bug: 188792659
Test: atest FlickerTests
Test: atest WMShellFlickerTests

+++ libraries/flicker/src/com/android/server/wm/traces/parser/layers/LayersTraceParser.kt
@@ -198,7 +198,7 @@ class LayersTraceParser {
 
         private fun newDisplay(proto: DisplayProto): Display {
             return Display(
-                proto.id,
+                proto.id.toULong(),
                 proto.name,
                 proto.layerStack,
                 proto.size.toSize(),
